Arreglo de warnings

+++ x64barebones/Kernel/include/sysCallDispatcher.h
@@ -7,6 +7,6 @@ typedef struct {
     uint64_t y;
 } coords;
 
-int sysCallDispatcher(uint64_t arg0, uint64_t arg1, uint64_t arg2, uint64_t arg3,
+uint64_t sysCallDispatcher(uint64_t arg0, uint64_t arg1, uint64_t arg2, uint64_t arg3,
                        uint64_t arg4, uint64_t arg5, uint64_t RAX);
 #endif //TPE_SYSCALLDISPATCHER_H
